+++ cmd/flexbot/main.go
@@ -19,7 +19,7 @@ import (
 )
 
 const (
-	version = "1.6.1"
+	version = "1.6.2"
 )
 
 type OperationResult interface {